server/agent/mcp/requirements.txt
fastapi>=0.103.0
uvicorn>=0.23.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pydantic>=1.10.0,<2.0.0
slack-sdk>=3.23.0
PyJWT>=2.8.0
cryptography>=41.0.0
requests>=2.31.0
python-multipart>=0.0.6
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.0
python-dotenv>=1.0.0 